// src/types/CitationInfoIDType.ts
export interface CitationInfoType {
  dateMartyrdom: string;
  burialDate: string;
  age: string;
  ageStatus: string;
  dissident: string;
  preRevolution: string;
  martyrdomGovernorate: string;
  countryOfMartyrdom: string;
  cityOfMartyrdom: string;
  martyrdomLocation: string;
  citationMethod: string;
  massacreId: string | null;
  massacre: string;
  sourceOfInformation?: string;
}